Add CoreWingMovement to _core.py
Extract shared WingMovement infrastructure into CoreWingMovement in
_core.py, following the same pattern used for
CoreWingCrossSectionMovement. Refactor WingMovement to extend it with an
empty __slots__ and a pass-through __init__. Use Sequence instead of
list for the wing_cross_section_movements parameter type on
CoreWingMovement to satisfy mypy covariance requirements.
