+++ pterasoftware/movements/wing_movement.py
@@ -11,17 +11,15 @@ None
 
 from __future__ import annotations
 
-import copy
 from collections.abc import Callable, Sequence
 
 import numpy as np
 
-from .. import _parameter_validation, _transformations, geometry
-from . import _functions
+from .. import _core, _parameter_validation, geometry
 from . import wing_cross_section_movement as wing_cross_section_movement_mod
 
 
-class WingMovement:
+class WingMovement(_core.CoreWingMovement):
     """A class used to contain a Wing's movement.
 
     **Contains the following methods:**
@@ -33,6 +31,8 @@ class WingMovement:
 
     max_period: The longest period of WingMovement's own motion and that of its sub
     movement objects.
+
+    generate_wing_at_time_step: Creates the Wing at a single time step.
 
     generate_wings: Creates the Wing at each time step, and returns them in a list.
 
@@ -46,21 +46,7 @@ class WingMovement:
     WingMovement's parent AirplaneMovement's parent Movement.
     """
 
-    __slots__ = (
-        "_base_wing",
-        "_wing_cross_section_movements",
-        "_ampLer_Gs_Cgs",
-        "_periodLer_Gs_Cgs",
-        "_spacingLer_Gs_Cgs",
-        "_phaseLer_Gs_Cgs",
-        "_ampAngles_Gs_to_Wn_ixyz",
-        "_periodAngles_Gs_to_Wn_ixyz",
-        "_spacingAngles_Gs_to_Wn_ixyz",
-        "_phaseAngles_Gs_to_Wn_ixyz",
-        "_rotationPointOffset_Gs_Ler",
-        "_all_periods",
-        "_max_period",
-    )
+    __slots__ = ()
 
     def __init__(
         self,
@@ -175,328 +161,19 @@ class WingMovement:
             occurs about the leading edge root point (default behavior). The units are
             in meters. The default is (0.0, 0.0, 0.0).
         """
-        # Validate and store immutable attributes. Set those that are numpy arrays to
-        # be read only.
-        if not isinstance(base_wing, geometry.wing.Wing):
-            raise TypeError("base_wing must be a Wing.")
-        self._base_wing = base_wing
-
-        if not isinstance(wing_cross_section_movements, list):
-            raise TypeError("wing_cross_section_movements must be a list.")
-        if len(wing_cross_section_movements) != len(
-            self._base_wing.wing_cross_sections
-        ):
-            raise ValueError(
-                "wing_cross_section_movements must have the same length as "
-                "base_wing.wing_cross_sections."
-            )
-        for wing_cross_section_movement in wing_cross_section_movements:
-            if not isinstance(
-                wing_cross_section_movement,
-                wing_cross_section_movement_mod.WingCrossSectionMovement,
-            ):
-                raise TypeError(
-                    "Every element in wing_cross_section_movements must be a "
-                    "WingCrossSectionMovement."
-                )
-        # Store as tuple to prevent external mutation.
-        self._wing_cross_section_movements = tuple(wing_cross_section_movements)
-
-        ampLer_Gs_Cgs = _parameter_validation.threeD_number_vectorLike_return_float(
-            ampLer_Gs_Cgs, "ampLer_Gs_Cgs"
+        super().__init__(
+            base_wing=base_wing,
+            wing_cross_section_movements=wing_cross_section_movements,
+            ampLer_Gs_Cgs=ampLer_Gs_Cgs,
+            periodLer_Gs_Cgs=periodLer_Gs_Cgs,
+            spacingLer_Gs_Cgs=spacingLer_Gs_Cgs,
+            phaseLer_Gs_Cgs=phaseLer_Gs_Cgs,
+            ampAngles_Gs_to_Wn_ixyz=ampAngles_Gs_to_Wn_ixyz,
+            periodAngles_Gs_to_Wn_ixyz=periodAngles_Gs_to_Wn_ixyz,
+            spacingAngles_Gs_to_Wn_ixyz=spacingAngles_Gs_to_Wn_ixyz,
+            phaseAngles_Gs_to_Wn_ixyz=phaseAngles_Gs_to_Wn_ixyz,
+            rotationPointOffset_Gs_Ler=rotationPointOffset_Gs_Ler,
         )
-        if not np.all(ampLer_Gs_Cgs >= 0.0):
-            raise ValueError("All elements in ampLer_Gs_Cgs must be non negative.")
-        self._ampLer_Gs_Cgs = ampLer_Gs_Cgs
-        self._ampLer_Gs_Cgs.flags.writeable = False
-
-        periodLer_Gs_Cgs = _parameter_validation.threeD_number_vectorLike_return_float(
-            periodLer_Gs_Cgs, "periodLer_Gs_Cgs"
-        )
-        if not np.all(periodLer_Gs_Cgs >= 0.0):
-            raise ValueError("All elements in periodLer_Gs_Cgs must be non negative.")
-        for period_index, period in enumerate(periodLer_Gs_Cgs):
-            amp = self._ampLer_Gs_Cgs[period_index]
-            if amp == 0 and period != 0:
-                raise ValueError(
-                    "If an element in ampLer_Gs_Cgs is 0.0, the corresponding element "
-                    "in periodLer_Gs_Cgs must be also be 0.0."
-                )
-        self._periodLer_Gs_Cgs = periodLer_Gs_Cgs
-        self._periodLer_Gs_Cgs.flags.writeable = False
-
-        # Store as tuple to prevent external mutation.
-        self._spacingLer_Gs_Cgs = (
-            _parameter_validation.threeD_spacing_vectorLike_return_tuple(
-                spacingLer_Gs_Cgs, "spacingLer_Gs_Cgs"
-            )
-        )
-
-        phaseLer_Gs_Cgs = _parameter_validation.threeD_number_vectorLike_return_float(
-            phaseLer_Gs_Cgs, "phaseLer_Gs_Cgs"
-        )
-        if not (np.all(phaseLer_Gs_Cgs > -180.0) and np.all(phaseLer_Gs_Cgs <= 180.0)):
-            raise ValueError(
-                "All elements in phaseLer_Gs_Cgs must be in the range (-180.0, 180.0]."
-            )
-        for phase_index, phase in enumerate(phaseLer_Gs_Cgs):
-            amp = self._ampLer_Gs_Cgs[phase_index]
-            if amp == 0 and phase != 0:
-                raise ValueError(
-                    "If an element in ampLer_Gs_Cgs is 0.0, the corresponding element "
-                    "in phaseLer_Gs_Cgs must be also be 0.0."
-                )
-        self._phaseLer_Gs_Cgs = phaseLer_Gs_Cgs
-        self._phaseLer_Gs_Cgs.flags.writeable = False
-
-        ampAngles_Gs_to_Wn_ixyz = (
-            _parameter_validation.threeD_number_vectorLike_return_float(
-                ampAngles_Gs_to_Wn_ixyz, "ampAngles_Gs_to_Wn_ixyz"
-            )
-        )
-        if not (
-            np.all(ampAngles_Gs_to_Wn_ixyz >= 0.0)
-            and np.all(ampAngles_Gs_to_Wn_ixyz <= 180.0)
-        ):
-            raise ValueError(
-                "All elements in ampAngles_Gs_to_Wn_ixyz must be in the range [0.0, "
-                "180.0]."
-            )
-        self._ampAngles_Gs_to_Wn_ixyz = ampAngles_Gs_to_Wn_ixyz
-        self._ampAngles_Gs_to_Wn_ixyz.flags.writeable = False
-
-        periodAngles_Gs_to_Wn_ixyz = (
-            _parameter_validation.threeD_number_vectorLike_return_float(
-                periodAngles_Gs_to_Wn_ixyz, "periodAngles_Gs_to_Wn_ixyz"
-            )
-        )
-        if not np.all(periodAngles_Gs_to_Wn_ixyz >= 0.0):
-            raise ValueError(
-                "All elements in periodAngles_Gs_to_Wn_ixyz must be non negative."
-            )
-        for period_index, period in enumerate(periodAngles_Gs_to_Wn_ixyz):
-            amp = self._ampAngles_Gs_to_Wn_ixyz[period_index]
-            if amp == 0 and period != 0:
-                raise ValueError(
-                    "If an element in ampAngles_Gs_to_Wn_ixyz is 0.0, "
-                    "the corresponding element in periodAngles_Gs_to_Wn_ixyz must be "
-                    "also be 0.0."
-                )
-        self._periodAngles_Gs_to_Wn_ixyz = periodAngles_Gs_to_Wn_ixyz
-        self._periodAngles_Gs_to_Wn_ixyz.flags.writeable = False
-
-        # Store as tuple to prevent external mutation.
-        self._spacingAngles_Gs_to_Wn_ixyz = (
-            _parameter_validation.threeD_spacing_vectorLike_return_tuple(
-                spacingAngles_Gs_to_Wn_ixyz, "spacingAngles_Gs_to_Wn_ixyz"
-            )
-        )
-
-        phaseAngles_Gs_to_Wn_ixyz = (
-            _parameter_validation.threeD_number_vectorLike_return_float(
-                phaseAngles_Gs_to_Wn_ixyz, "phaseAngles_Gs_to_Wn_ixyz"
-            )
-        )
-        if not (
-            np.all(phaseAngles_Gs_to_Wn_ixyz > -180.0)
-            and np.all(phaseAngles_Gs_to_Wn_ixyz <= 180.0)
-        ):
-            raise ValueError(
-                "All elements in phaseAngles_Gs_to_Wn_ixyz must be in the range ("
-                "-180.0, 180.0]."
-            )
-        for phase_index, phase in enumerate(phaseAngles_Gs_to_Wn_ixyz):
-            amp = self._ampAngles_Gs_to_Wn_ixyz[phase_index]
-            if amp == 0 and phase != 0:
-                raise ValueError(
-                    "If an element in ampAngles_Gs_to_Wn_ixyz is 0.0, "
-                    "the corresponding element in phaseAngles_Gs_to_Wn_ixyz must be "
-                    "also be 0.0."
-                )
-        self._phaseAngles_Gs_to_Wn_ixyz = phaseAngles_Gs_to_Wn_ixyz
-        self._phaseAngles_Gs_to_Wn_ixyz.flags.writeable = False
-
-        rotationPointOffset_Gs_Ler = (
-            _parameter_validation.threeD_number_vectorLike_return_float(
-                rotationPointOffset_Gs_Ler, "rotationPointOffset_Gs_Ler"
-            )
-        )
-        self._rotationPointOffset_Gs_Ler = rotationPointOffset_Gs_Ler
-        self._rotationPointOffset_Gs_Ler.flags.writeable = False
-
-        # Initialize the caches for the properties derived from the immutable
-        # attributes.
-        self._all_periods: tuple[float, ...] | None = None
-        self._max_period: float | None = None
-
-    # --- Deep copy method ---
-    def __deepcopy__(self, memo: dict) -> WingMovement:
-        """Creates a deep copy of this WingMovement.
-
-        All attributes are copied. The base Wing and WingCrossSectionMovements are deep
-        copied to ensure independence. NumPy arrays are copied and set to read only to
-        preserve immutability. Cache variables are reset to None.
-
-        :param memo: A dict used by the copy module to track already copied objects and
-            avoid infinite recursion.
-        :return: A new WingMovement with copied attributes.
-        """
-        # Create a new WingMovement instance without calling __init__ to avoid
-        # redundant validation.
-        new_movement = object.__new__(WingMovement)
-
-        # Store this WingMovement in memo to handle potential circular references.
-        memo[id(self)] = new_movement
-
-        # Deep copy the base Wing to ensure independence (immutable).
-        new_movement._base_wing = copy.deepcopy(self._base_wing, memo)
-
-        # Deep copy the WingCrossSectionMovements and store as tuple.
-        new_movement._wing_cross_section_movements = tuple(
-            copy.deepcopy(wing_cross_section_movement, memo)
-            for wing_cross_section_movement in self._wing_cross_section_movements
-        )
-
-        # Copy numpy arrays and make them read only.
-        new_movement._ampLer_Gs_Cgs = self._ampLer_Gs_Cgs.copy()
-        new_movement._ampLer_Gs_Cgs.flags.writeable = False
-
-        new_movement._periodLer_Gs_Cgs = self._periodLer_Gs_Cgs.copy()
-        new_movement._periodLer_Gs_Cgs.flags.writeable = False
-
-        new_movement._phaseLer_Gs_Cgs = self._phaseLer_Gs_Cgs.copy()
-        new_movement._phaseLer_Gs_Cgs.flags.writeable = False
-
-        new_movement._ampAngles_Gs_to_Wn_ixyz = self._ampAngles_Gs_to_Wn_ixyz.copy()
-        new_movement._ampAngles_Gs_to_Wn_ixyz.flags.writeable = False
-
-        new_movement._periodAngles_Gs_to_Wn_ixyz = (
-            self._periodAngles_Gs_to_Wn_ixyz.copy()
-        )
-        new_movement._periodAngles_Gs_to_Wn_ixyz.flags.writeable = False
-
-        new_movement._phaseAngles_Gs_to_Wn_ixyz = self._phaseAngles_Gs_to_Wn_ixyz.copy()
-        new_movement._phaseAngles_Gs_to_Wn_ixyz.flags.writeable = False
-
-        new_movement._rotationPointOffset_Gs_Ler = (
-            self._rotationPointOffset_Gs_Ler.copy()
-        )
-        new_movement._rotationPointOffset_Gs_Ler.flags.writeable = False
-
-        # Copy tuples directly (they are immutable).
-        new_movement._spacingLer_Gs_Cgs = self._spacingLer_Gs_Cgs
-        new_movement._spacingAngles_Gs_to_Wn_ixyz = self._spacingAngles_Gs_to_Wn_ixyz
-
-        # Initialize cache variables to None (caches will be recomputed on access).
-        new_movement._all_periods = None
-        new_movement._max_period = None
-
-        return new_movement
-
-    # --- Immutable: read only properties ---
-    @property
-    def base_wing(self) -> geometry.wing.Wing:
-        return self._base_wing
-
-    @property
-    def wing_cross_section_movements(
-        self,
-    ) -> tuple[wing_cross_section_movement_mod.WingCrossSectionMovement, ...]:
-        return self._wing_cross_section_movements
-
-    @property
-    def ampLer_Gs_Cgs(self) -> np.ndarray:
-        return self._ampLer_Gs_Cgs
-
-    @property
-    def periodLer_Gs_Cgs(self) -> np.ndarray:
-        return self._periodLer_Gs_Cgs
-
-    @property
-    def spacingLer_Gs_Cgs(
-        self,
-    ) -> tuple[str | Callable[[float], float], ...]:
-        return self._spacingLer_Gs_Cgs
-
-    @property
-    def phaseLer_Gs_Cgs(self) -> np.ndarray:
-        return self._phaseLer_Gs_Cgs
-
-    @property
-    def ampAngles_Gs_to_Wn_ixyz(self) -> np.ndarray:
-        return self._ampAngles_Gs_to_Wn_ixyz
-
-    @property
-    def periodAngles_Gs_to_Wn_ixyz(self) -> np.ndarray:
-        return self._periodAngles_Gs_to_Wn_ixyz
-
-    @property
-    def spacingAngles_Gs_to_Wn_ixyz(
-        self,
-    ) -> tuple[str | Callable[[float], float], ...]:
-        return self._spacingAngles_Gs_to_Wn_ixyz
-
-    @property
-    def phaseAngles_Gs_to_Wn_ixyz(self) -> np.ndarray:
-        return self._phaseAngles_Gs_to_Wn_ixyz
-
-    @property
-    def rotationPointOffset_Gs_Ler(self) -> np.ndarray:
-        return self._rotationPointOffset_Gs_Ler
-
-    # --- Immutable derived: manual lazy caching ---
-    @property
-    def all_periods(self) -> tuple[float, ...]:
-        """All unique non zero periods from this WingMovement and its
-        WingCrossSectionMovements.
-
-        :return: A tuple of all unique non zero periods in seconds. If all motion is
-            static, this will be an empty tuple.
-        """
-        if self._all_periods is None:
-            periods: list[float] = []
-
-            # Collect all periods from WingCrossSectionMovements.
-            for wing_cross_section_movement in self._wing_cross_section_movements:
-                periods.extend(wing_cross_section_movement.all_periods)
-
-            # Collect all periods from WingMovement's own motion.
-            for period in self._periodLer_Gs_Cgs:
-                if period > 0.0:
-                    periods.append(float(period))
-            for period in self._periodAngles_Gs_to_Wn_ixyz:
-                if period > 0.0:
-                    periods.append(float(period))
-
-            self._all_periods = tuple(periods)
-        return self._all_periods
-
-    @property
-    def max_period(self) -> float:
-        """The longest period of WingMovement's own motion and that of its sub movement
-        objects.
-
-        :return: The longest period in seconds. If all the motion is static, this will
-            be 0.0.
-        """
-        if self._max_period is None:
-            wing_cross_section_movement_max_periods = []
-            for wing_cross_section_movement in self._wing_cross_section_movements:
-                wing_cross_section_movement_max_periods.append(
-                    wing_cross_section_movement.max_period
-                )
-            max_wing_cross_section_movement_period = max(
-                wing_cross_section_movement_max_periods
-            )
-
-            self._max_period = float(
-                max(
-                    max_wing_cross_section_movement_period,
-                    np.max(self._periodLer_Gs_Cgs),
-                    np.max(self._periodAngles_Gs_to_Wn_ixyz),
-                )
-            )
-        return self._max_period
 
     # --- Other methods ---
     def generate_wings(
@@ -521,195 +198,8 @@ class WingMovement:
             delta_time, "delta_time", min_val=0.0, min_inclusive=False
         )
 
-        # Get the time at each time step.
-        times = np.linspace(
-            0.0, num_steps * delta_time, num_steps, endpoint=False, dtype=float
-        )
-
-        # Generate oscillating values for each dimension of Ler_Gs_Cgs.
-        listLer_Gs_Cgs = np.zeros((3, num_steps), dtype=float)
-        for dim in range(3):
-            this_spacing = self._spacingLer_Gs_Cgs[dim]
-            this_amp = self._ampLer_Gs_Cgs[dim]
-            this_period = self._periodLer_Gs_Cgs[dim]
-            this_phase = self._phaseLer_Gs_Cgs[dim]
-            this_base = self._base_wing.Ler_Gs_Cgs[dim]
-
-            if this_spacing == "sine":
-                for this_time_step, this_time in enumerate(times):
-                    listLer_Gs_Cgs[dim, this_time_step] = (
-                        _functions.oscillating_sin_at_time(
-                            amp=this_amp,
-                            period=this_period,
-                            phase=this_phase,
-                            base=this_base,
-                            time=this_time,
-                        )
-                    )
-            elif this_spacing == "uniform":
-                for this_time_step, this_time in enumerate(times):
-                    listLer_Gs_Cgs[dim, this_time_step] = (
-                        _functions.oscillating_lin_at_time(
-                            amp=this_amp,
-                            period=this_period,
-                            phase=this_phase,
-                            base=this_base,
-                            time=this_time,
-                        )
-                    )
-            elif callable(this_spacing):
-                for this_time_step, this_time in enumerate(times):
-                    listLer_Gs_Cgs[dim, this_time_step] = (
-                        _functions.oscillating_custom_at_time(
-                            amp=this_amp,
-                            period=this_period,
-                            phase=this_phase,
-                            base=this_base,
-                            time=this_time,
-                            custom_function=this_spacing,
-                        )
-                    )
-            else:
-                raise ValueError(f"Invalid spacing value: {this_spacing}")
-
-        # Generate oscillating values for each dimension of angles_Gs_to_Wn_ixyz.
-        listAngles_Gs_to_Wn_ixyz = np.zeros((3, num_steps), dtype=float)
-        for dim in range(3):
-            this_spacing = self._spacingAngles_Gs_to_Wn_ixyz[dim]
-            this_amp = self._ampAngles_Gs_to_Wn_ixyz[dim]
-            this_period = self._periodAngles_Gs_to_Wn_ixyz[dim]
-            this_phase = self._phaseAngles_Gs_to_Wn_ixyz[dim]
-            this_base = self._base_wing.angles_Gs_to_Wn_ixyz[dim]
-
-            if this_spacing == "sine":
-                for this_time_step, this_time in enumerate(times):
-                    listAngles_Gs_to_Wn_ixyz[dim, this_time_step] = (
-                        _functions.oscillating_sin_at_time(
-                            amp=this_amp,
-                            period=this_period,
-                            phase=this_phase,
-                            base=this_base,
-                            time=this_time,
-                        )
-                    )
-            elif this_spacing == "uniform":
-                for this_time_step, this_time in enumerate(times):
-                    listAngles_Gs_to_Wn_ixyz[dim, this_time_step] = (
-                        _functions.oscillating_lin_at_time(
-                            amp=this_amp,
-                            period=this_period,
-                            phase=this_phase,
-                            base=this_base,
-                            time=this_time,
-                        )
-                    )
-            elif callable(this_spacing):
-                for this_time_step, this_time in enumerate(times):
-                    listAngles_Gs_to_Wn_ixyz[dim, this_time_step] = (
-                        _functions.oscillating_custom_at_time(
-                            amp=this_amp,
-                            period=this_period,
-                            phase=this_phase,
-                            base=this_base,
-                            time=this_time,
-                            custom_function=this_spacing,
-                        )
-                    )
-            else:
-                raise ValueError(f"Invalid spacing value: {this_spacing}")
-
-        # Create an empty 2D ndarray that will hold each of the Wings's
-        # WingCrossSection's vector of WingCrossSections representing its changing
-        # state at each time step. The first index denotes a particular base
-        # WingCrossSection, and the second index denotes the time step.
-        wing_cross_sections = np.empty(
-            (len(self._wing_cross_section_movements), num_steps), dtype=object
-        )
-
-        # Iterate through the WingCrossSectionMovements.
-        for (
-            wing_cross_section_movement_id,
-            wing_cross_section_movement,
-        ) in enumerate(self._wing_cross_section_movements):
-
-            # Generate this WingCrossSection's vector of WingCrossSections
-            # representing its changing state at each time step.
-            this_wing_cross_sections_list_of_wing_cross_sections = np.array(
-                wing_cross_section_movement.generate_wing_cross_sections(
-                    num_steps=num_steps, delta_time=delta_time
-                )
-            )
-
-            # Add this vector the Wing's 2D ndarray of WingCrossSections'
-            # WingCrossSections.
-            wing_cross_sections[wing_cross_section_movement_id, :] = (
-                this_wing_cross_sections_list_of_wing_cross_sections
-            )
-
-        # Create an empty list to hold each time step's Wing.
         wings = []
-
-        # Get the non changing Wing attributes.
-        this_name = self._base_wing.name
-        this_symmetric = self._base_wing.symmetric
-        this_mirror_only = self._base_wing.mirror_only
-        this_symmetryNormal_G = self._base_wing.symmetryNormal_G
-        this_symmetryPoint_G_Cg = self._base_wing.symmetryPoint_G_Cg
-        this_num_chordwise_panels = self._base_wing.num_chordwise_panels
-        this_chordwise_spacing = self._base_wing.chordwise_spacing
-
-        # Check if there is any offset rotation.
-        offset_rotation = not np.allclose(
-            self._rotationPointOffset_Gs_Ler, np.zeros(3, dtype=float)
-        )
-
-        # Initialize an identity matrix.
-        identity = np.eye(3, dtype=float)
-
-        # Iterate through the time steps.
         for step in range(num_steps):
-            thisLer_Gs_Cgs = listLer_Gs_Cgs[:, step]
-            theseAngles_Gs_to_Wn_ixyz = listAngles_Gs_to_Wn_ixyz[:, step]
-            these_wing_cross_sections = list(wing_cross_sections[:, step])
-
-            # If there is a non zero rotation point offset, adjust the positions to
-            # account for rotation about the offset point instead of the leading edge
-            # root.
-            if offset_rotation:
-                # TODO: Refactor this procedure for producing offset rotations to be a
-                #  function in _transformations.py.
-                # Get the active rotation matrix for this step's angles.
-                rot_T_act = _transformations.generate_rot_T(
-                    theseAngles_Gs_to_Wn_ixyz,
-                    passive=False,
-                    intrinsic=True,
-                    order="xyz",
-                )
-                rot_R_act = rot_T_act[:3, :3]
-
-                # Compute the position adjustment due to the offset rotation point.
-                offsetRotationPointAdjustment_Gs = (
-                    identity - rot_R_act
-                ) @ self._rotationPointOffset_Gs_Ler
-
-                # Apply the position adjustment to the leading edge root.
-                thisLer_Gs_Cgs = thisLer_Gs_Cgs + offsetRotationPointAdjustment_Gs
-
-            # Make a new Wing for this time step.
-            this_wing = geometry.wing.Wing(
-                wing_cross_sections=these_wing_cross_sections,
-                name=this_name,
-                Ler_Gs_Cgs=thisLer_Gs_Cgs,
-                angles_Gs_to_Wn_ixyz=theseAngles_Gs_to_Wn_ixyz,
-                symmetric=this_symmetric,
-                mirror_only=this_mirror_only,
-                symmetryNormal_G=this_symmetryNormal_G,
-                symmetryPoint_G_Cg=this_symmetryPoint_G_Cg,
-                num_chordwise_panels=this_num_chordwise_panels,
-                chordwise_spacing=this_chordwise_spacing,
-            )
-
-            # Add this new Wing to the list of Wings.
-            wings.append(this_wing)
+            wings.append(self.generate_wing_at_time_step(step, delta_time))
 
         return wings
